requirements.txt
fastapi
torch
fastapi-swagger-dark
pydantic
sentence-transformers
pydantic-settings
python-dotenv
uvicorn
ruff
pytest
pytest-cov
httpx
pytest-mock
llama-cpp-python
python-multipart
